Feature/crud

+++ src/main/java/com/etna/gpe/controller/OrganizationController.java
@@ -1,8 +1,10 @@
 package com.etna.gpe.controller;
 
+import com.etna.gpe.controller.customexception.ParametersNotFound;
 import com.etna.gpe.dto.OrganizationDto;
 import com.etna.gpe.service.OrganizationService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -15,24 +17,33 @@ public class OrganizationController {
     OrganizationService organizationService;
 
     @GetMapping("/all_organization")
+    @ResponseStatus(HttpStatus.OK)
     List<OrganizationDto> getAllOrganization() {
         return organizationService.getAllOrganization();
     }
 
     @PostMapping("/create_organization")
-    void createOrganization(@RequestBody OrganizationDto organizationDto) {
-        organizationService.createOrUpdateuOrganization(organizationDto);
+    @ResponseStatus(HttpStatus.CREATED)
+    OrganizationDto createOrganization(@RequestBody OrganizationDto organizationDto) {
+    	if(organizationDto == null)
+    		throw new ParametersNotFound();
+        return organizationService.createOrUpdateuOrganization(organizationDto);
     }
 
     @GetMapping("/get_organization")
-    OrganizationDto getParticularByEmail(@RequestParam(value = "email") String email) {
-        return
-                organizationService.getOrganizationByEmail(email);
+    @ResponseStatus(HttpStatus.OK)
+    OrganizationDto getOrganizationByEmail(@RequestParam(value = "email") String email) {
+    	if(email.isEmpty() || email == null)
+    		throw new ParametersNotFound();
+        return organizationService.getOrganizationByEmail(email, false);
     }
 
+    //todo post a requestbody and not delete really but put boolean to true
     @PostMapping("delete_organization")
-    void
-    deleteparticular(@RequestParam(value = "email") String email) {
+    @ResponseStatus(HttpStatus.RESET_CONTENT)
+    void deleteparticular(@RequestParam(value = "email") String email) {
+    	if(email.isEmpty() || email == null)
+    		throw new ParametersNotFound();
         organizationService.deleteOrganization(email);
     }
