+++ src/main/java/com/etna/gpe/controller/ParticularController.java
@@ -1,8 +1,10 @@
 package com.etna.gpe.controller;
 
+import com.etna.gpe.controller.customexception.ParametersNotFound;
 import com.etna.gpe.dto.ParticularDto;
 import com.etna.gpe.service.ParticularService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -11,32 +13,38 @@ import java.util.List;
 @RequestMapping("/particular")
 public class ParticularController {
 
-    @Autowired
-    ParticularService particularService;
+	@Autowired
+	ParticularService particularService;
 
-    @GetMapping("/all_particular")
-    List<ParticularDto> getAllParticular() {
-        return particularService.getAllParticular();
-    }
+	@GetMapping("/all_particular")
+	@ResponseStatus(HttpStatus.OK)
+	List<ParticularDto> getAllParticular() {
+		return particularService.getAllParticular();
+	}
 
-    @PostMapping("/create_particular")
-    void createParticular(@RequestBody ParticularDto particularDto) {
-        particularService.createOrUpdateuParticular(particularDto.getParticularEmail(
-                ), particularDto.getParticularPassword(), particularDto.getParticularName(),
-                particularDto.getParticularFirstName(),
-                particularDto.getParticularPhonenumber());
-    }
+	@PostMapping("/create_particular")
+	@ResponseStatus(HttpStatus.CREATED)
+	ParticularDto createParticular(@RequestBody ParticularDto particularDto) {
+		if (particularDto == null)
+			throw new ParametersNotFound();
+		return particularService.createOrUpdateuParticular(particularDto);
+	}
 
-    @GetMapping("/get_particular")
-    ParticularDto getParticularByEmail(@RequestParam(value = "email") String email) {
-        return
-                particularService.getParticularByEmail(email);
-    }
+	@GetMapping("/get_particular")
+	@ResponseStatus(HttpStatus.OK)
+	ParticularDto getParticularByEmail(@RequestParam(value = "email") String email) {
+		if (email == null || email.isEmpty())
+			throw new ParametersNotFound();
+		return particularService.getParticularByEmail(email);
+	}
 
-    @PostMapping("/delete_particulier")
-    void deleteparticular(@RequestParam(value = "email") String email) {
-        particularService.deleteParticular(email);
-    }
-
+	// todo post a requestbodyn and not delete really but put boolean to true
+	@PostMapping("/delete_particulier")
+	@ResponseStatus(HttpStatus.RESET_CONTENT)
+	void deleteparticular(@RequestParam(value = "email") String email) {
+		if (email == null || email.isEmpty())
+			throw new ParametersNotFound();
+		particularService.deleteParticular(email);
+	}
 
 }
